+++ sct_sentinel1_reader/src/sct_sentinel1_reader/protocol_implementation.py
@@ -156,6 +156,9 @@ class Sentinel1ChannelManager:
         self._metadata_file = channel_metadata_path
         self._raster_file = channel_raster_path
         self._channel = read_channel_metadata(xml_path=channel_metadata_path, external_orbit_path=external_orbit_path)
+        self._sensor_name = (
+            "" if self._channel.dataset_info.sensor_name is None else self._channel.dataset_info.sensor_name
+        )
 
         # translating arepyextras.eo_products enum to arepyextras.quality ones
         # NOTE: BETA NOUGHT fixed and taking only first value of first calibration vector
@@ -336,6 +339,11 @@ class Sentinel1ChannelManager:
             ]
 
         return burst_az_boundaries, burst_rng_boundaries
+
+    @property
+    def sensor_name(self) -> str:
+        """Name of the sensor"""
+        return self._sensor_name
 
     @property
     def swath_name(self) -> str: